Started the redesign process for the character-create page
At the moment I have an error regarding the useMemo hook. Need some assistance here.

Also I have an issue with images because i can't render anything inside the <img> tag but from scss it renders ok. Also i tried to modularize as much and I thought to include the description and the sprite for each class inside the json file. Any thoughts on that?